Added authorization to posts - REady for live deployment

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,11 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['getActivate', 'index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -44,6 +49,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -70,6 +76,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'You are not authorized to edit that post');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -104,6 +115,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'You are not authorized to edit that post');
+        }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
